+++ tests/models.py
@@ -2,7 +2,14 @@
 
 from __future__ import annotations
 
-from sparqlmodel import Field, Relationship, SPARQLModel
+from sparqlmodel import IRI, Field, Relationship, SPARQLModel
+
+
+class Location(SPARQLModel):
+    rdf_type = "schema:Place"
+    __prefixes__ = {"schema": "https://schema.org/"}
+
+    name: str = Field("schema:name")
 
 
 class Organization(SPARQLModel):
@@ -10,6 +17,7 @@ class Organization(SPARQLModel):
     __prefixes__ = {"schema": "https://schema.org/"}
 
     name: str = Field("schema:name")
+    located_in: Location | None = Relationship("schema:location", model=Location)
 
 
 class Person(SPARQLModel):
@@ -17,4 +25,4 @@ class Person(SPARQLModel):
     __prefixes__ = {"schema": "https://schema.org/"}
 
     name: str = Field("schema:name")
-    works_for: Organization | None = Relationship("schema:worksFor", model=Organization)
+    works_for: Organization | IRI | None = Relationship("schema:worksFor", model=Organization)